First Version of Di-Finder
Note:
The steps of the tool are numbered from 1 to 3.
The get sequence process can be time consuming (Please wait until sequences download).
After receiving the sequences, they will be saved in the appropriate folder.
The last step ends with finding the STR and you can save the results by clicking the save result button.